Split out LocalScriptRunner
